+++ src/main/webapp/WEB-INF/pages/tasks.jsp
@@ -1,4 +1,4 @@
-<%@ page import="by.bsuir.bugTrackingSystem.model.Employee" %>
+<%@ page import="by.a1qa.model.Employee" %>
 <%@ taglib uri="http://java.sun.com/jsp/jstl/core" prefix="c" %>
 <%@ taglib uri="http://www.springframework.org/tags" prefix="spring" %>
 <%@ taglib uri="http://www.springframework.org/tags/form" prefix="form" %>
